refactor(schema-api): extract TableApi and IndexApi traits
Extract table and index operations from SchemaApi into separate
traits to improve modularity and separation of concerns. This
change organizes the codebase into focused, specialized traits
for better maintainability and clearer API boundaries.

Key changes:
- Create TableApi trait in src/meta/api/src/table_api.rs with all
  table lifecycle operations (create, drop, rename, truncate, etc.)
- Create IndexApi trait in src/meta/api/src/index_api.rs with all
  index management operations (create, drop, list, gc operations)
- Move corresponding implementations from schema_api.rs to the new
  trait files while preserving all existing functionality
- Update imports across the codebase to use the new traits
- Add required type parameters for generic method calls where the
  Rust compiler cannot infer types automatically
- Make internal utility functions public where needed by the new
  trait implementations

This refactoring maintains full API compatibility while creating
a cleaner, more organized code structure that will facilitate
future development and testing.
