+++ src/meta/api/src/lib.rs
@@ -22,6 +22,7 @@ mod data_mask_api;
 mod data_mask_api_impl;
 mod database_api;
 pub mod database_util;
+pub mod index_api;
 pub mod kv_app_error;
 pub mod kv_pb_api;
 pub mod kv_pb_crud_api;
@@ -32,6 +33,7 @@ pub mod reply;
 mod schema_api;
 mod schema_api_test_suite;
 mod sequence_api;
+pub mod table_api;
 pub(crate) mod testing;
 pub mod txn_backoff;
 pub mod util;
@@ -45,10 +47,12 @@ pub(crate) mod sequence_nextval_impl;
 
 pub use data_mask_api::DatamaskApi;
 pub use database_api::DatabaseApi;
+pub use index_api::IndexApi;
 pub use row_access_policy_api::RowAccessPolicyApi;
 pub use schema_api::SchemaApi;
 pub use schema_api_test_suite::SchemaApiTestSuite;
 pub use sequence_api::SequenceApi;
+pub use table_api::TableApi;
 pub use util::assert_table_exist;
 pub use util::db_has_to_exist;
 pub use util::deserialize_struct;
