+++ src/meta/binaries/metabench/main.rs
@@ -28,7 +28,7 @@ use databend_common_base::runtime;
 use databend_common_meta_api::serialize_struct;
 use databend_common_meta_api::txn_op_put;
 use databend_common_meta_api::DatabaseApi;
-use databend_common_meta_api::SchemaApi;
+use databend_common_meta_api::TableApi;
 use databend_common_meta_app::schema::database_name_ident::DatabaseNameIdent;
 use databend_common_meta_app::schema::CreateDatabaseReq;
 use databend_common_meta_app::schema::CreateOption;
