+++ src/query/service/src/catalogs/default/mutable_catalog.rs
@@ -26,8 +26,10 @@ use databend_common_exception::Result;
 use databend_common_meta_api::kv_app_error::KVAppError;
 use databend_common_meta_api::name_id_value_api::NameIdValueApiCompat;
 use databend_common_meta_api::DatabaseApi;
+use databend_common_meta_api::IndexApi;
 use databend_common_meta_api::SchemaApi;
 use databend_common_meta_api::SequenceApi;
+use databend_common_meta_api::TableApi;
 use databend_common_meta_app::app_error::AppError;
 use databend_common_meta_app::schema::database_name_ident::DatabaseNameIdent;
 use databend_common_meta_app::schema::dictionary_name_ident::DictionaryNameIdent;
